Add channel closing
